Refactor#42/MemberRepository-Refactoring
