+++ src/test/java/com/sungkyul/decemberproject/Awesomely_Delicious/RestDocsConfiguration.java
@@ -1,18 +1,10 @@
 package com.sungkyul.decemberproject.Awesomely_Delicious;
 
-import antlr.preprocessor.Preprocessor;
-import com.sungkyul.decemberproject.Awesomely_Delicious.repository.MemberRepository;
-import com.sungkyul.decemberproject.Awesomely_Delicious.repository.RestaurantRepository;
-import com.sungkyul.decemberproject.Awesomely_Delicious.service.RestaurantService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
 import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
 import org.springframework.restdocs.operation.preprocess.Preprocessors;
-
-import javax.persistence.EntityManager;
 
 @TestConfiguration
 public class RestDocsConfiguration {
